requirements.txt
colorama==0.4.6
playwright==1.46.0
pyperclip==1.9.0
python-decouple==3.8
requests~=2.32.3
selenium==4.24.0
webdriver-manager==4.0.2